Backend/requirements.txt
fastapi
Pillow
uvicorn
pydantic
google.generativeai
python-dotenv